requirements.txt
gensim==4.0.1
jieba==0.42.1
numpy==1.17.4
pandas==1.1.3
Pillow==9.1.0
prefetch_generator==1.0.1
pytorch_warmup==0.0.4
requests==2.26.0
torch==1.10.0+cu102
torchvision==0.9.1
tqdm==4.41.0
transformers==4.11.3
vit_pytorch==0.27.0
